requirements.txt
# This file is only for local development / testing.
# It is not used by Home Assistant.

PyYAML~=6.0.1 # Note: this is not mentioned in havs.json as it's already available in home assistant
pymodbus~=3.6.2
homeassistant~=2023.12.4
pytest~=7.4.3
pytest-cov~=4.1.0
pytest-homeassistant-custom-component~=0.13.85
pytest-xdist~=3.5.0

# todo: missing types
